Add initial "session" route to handle setting of Cookies
The last step of the OAuth dance required some sort of session store in order
to store the initial set of OAuth tokens. Because this application does not
use any sort of data store, the only option available is to use cookies. But
because the 2nd step of the OAuth dance takes place on JIRA's servers, we
cannot issue a Set-Cookie header on their domain and read it later. So before
redirected to JIRA, the app redirects internally to this "session" route whose
sole purpose is to set the cookie and then redirect elsewhere. That way, when
JIRA points back to this application, the OAuth tokens can be retrieved from
the cookie and used to generate the access token.