Simplify RSA Key integration by utilizing S3
This adds the RSA Keys to an S3 bucket (`ori-rsa-keygen`) so that they can be
pulled out programmatically from within the lambda function. This is all
encapsulated by the `oauthConsumer` function which is part of `ori-tools` and
is shared amongst the other lambda functions.

*   `oauthConsumer` pulls the keys from S3 and creates an OAuth consumer that
can be used to sign requests to OAuth APIs, specifically JIRA.

*   Key generation and syncing is now handled entirely by `ori-rsa-keygen`
and is no longer intended to be used by other functions. This fully
encapsulates key generation and syncing with S3. See the README for more info.